add state to milestone
functionality added:

* signal to reviewer that the milestone is ready for a review
* allow reviewer to reject that the milestone is completed
* emit StateChanged events

The event allows us to efficently subscribe to changes across all
lpp-milestone contracts. We can create a filter in with the topics array
of:
[
  hash_of_StateChanged_event_signature,
  64_byte_padded_address_of_liquidPledging_contract
]

This solves the challenge of how to keep track of all the milestone
contract current states. The above topics array allows us to react to
any changes in a milestone for the liquidPledging contract we are
interested in.