remove canceled bool from contract
This fixes a minor bug where the lpp-milestone contract canceled bool
doesn't accurately reflect the current status. If a parent project is
canceled in liquidPledging, then the child projects are canceled as
well. This would result in the lpp-milestone being canceled in
liquidPledging, but the canceled bool == False in the contract.

replaced with a call to liquidPledging.isCanceled. This is to ensure
that the contract is always in the correct state.